requirements.txt
# Core Django
Django==5.0.1
djangorestframework==3.14.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Security
bcrypt==4.1.2
PyJWT==2.8.0
djangorestframework-simplejwt==5.3.1

# Environment Variables
python-decouple==3.8

# CORS
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.7

# Validation
email-validator==2.1.0.post1

# Rate Limiting
django-ratelimit==4.1.0

# Testing
pytest==7.4.4
pytest-django==4.7.0
factory-boy==3.3.0

# Development
django-extensions==3.2.3
ipython==8.20.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# Celery (para tarefas em segundo plano)
celery[redis]==5.3.6